requirements.txt
easygui==0.98.3
keyboard==0.13.5
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
pyflakes==2.4.0
PyGetWindow==0.0.9
PyRect==0.2.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
six==1.16.0